remove login activity from worker branch move WorkerDashboardActivity to the activities package
