--- app/src/main/java/com/ashstudios/safana/LoginActivity.java
@@ -1,19 +1,0 @@
-package com.ashstudios.safana;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Context;
-import android.content.Intent;
-import android.os.Bundle;
-
-public class LoginActivity extends AppCompatActivity {
-    private Context context;
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_login);
-        context = LoginActivity.this;
-        Intent intent = new Intent(context,WorkerDashboardActivity.class);
-        startActivity(intent);
-    }
-}
